Update README, add new features and improve code efficiency
This commit involves the addition of a 'Versions' section in the README.md displaying SDK and PHP versions used, corrections in the order of entities in the implementation of 'Card' strings, and adjustments in the transactions service. It also introduces a new feature that allows users to add new cards, with essential parameters being the card balance, card status, and card percentage. Efficiency in code has been improved through the replacement of several for loops with the more efficient list method, and cleaner code has been achieved through corrections in punctuations in the README.md file. All changes have been documented in the changelog.