Refactor Credentials and Context classes
The Credentials class has been refactored to provide readonly access to attributes. For the Context class, it has been changed from an enum to a class with additional validation in its constructor and methods to check the type of context. It now includes methods to verify whether the context is user, admin, or default.

Signed-off-by: B24io <app@b24.io>